fix comment

+++ YesterdayNews/Utils/NewsCacheUtils/NewsCacheDB.m
@@ -17,7 +17,7 @@
     return [[IDBManager getInstance] execSQL:SQL];
 }
 
-+ (BOOL)updateNewsComments:(int)comments WithID:(NSString *)newsID {
++ (BOOL)updateNewsComments:(NSInteger)comments WithID:(NSString *)newsID {
     NSString *SQL = [[NSString alloc] initWithFormat:@"UPDATE NewsCache SET comments = '%d' WHERE ID = '%@';", comments, newsID];
     return [[IDBManager getInstance] execSQL:SQL];
 }
